Update README.md
Udiddit is a social news aggregator, content rating, and discussions website. It allows registered users are able to link to external content or post their own text content about various topics, ranging from common topics such as food to more arcane ones such as coding. Other users can comment on these posts, and each user is allowed to cast a vote about each post, either in an up (like) or down (dislike) direction. Unfortunately, due to some time constraints before the initial launch of the site, the data model stored in Postgres hasn’t been well thought out, and is starting to show its flaws. An assessment of the situation was needed and steps taken to fix all the issues with the current data model. Once successful, an improvement of the current system was needed to make it more robust and also adding some web analytics.